fix: condense keydownAction to single file
