+++ src/pages/home/report/ReportActionCompose/keyDownAction.js
@@ -1,0 +1,10 @@
+function listenKeyDown(callback) {
+    document.addEventListener('keydown', callback);
+}
+
+function removeListenKeyDown(callback) {
+    document.removeEventListener('keydown', callback);
+}
+
+export default {listenKeyDown, removeListenKeyDown};
+
