--- src/pages/home/report/ReportActionCompose/listenKeyDown.js
@@ -1,5 +1,0 @@
-function listenKeyDown(callback) {
-    document.addEventListener('keydown', callback);
-}
-
-export default listenKeyDown;
