--- src/pages/home/report/ReportActionCompose/listenKeyDown.native.js
@@ -1,3 +1,0 @@
-function listenKeyDown() {}
-
-export default listenKeyDown;
