--- src/pages/home/report/ReportActionCompose/removeListenKeyDown.js
@@ -1,5 +1,0 @@
-function removeListenKeyDown(callback) {
-    document.removeEventListener('keydown', callback);
-}
-
-export default removeListenKeyDown;
