--- src/pages/home/report/ReportActionCompose/removeListenKeyDown.native.js
@@ -1,3 +1,0 @@
-function removeListenKeyDown() {}
-
-export default removeListenKeyDown;
